See description
Removed recusive layer parameter for best child, it was causing problems and wasn't doing what I wanted anyway. That can be fixed later. NExt fixed some bugs and caused some bugs in the IS_complete(). MOST of the time it correctly attributes the winner to either X/O/TIE but there is a bug in which the game isnt marked as complete but there are no legal moves. In this case I made it count as a tie.